Add missing visibility for methods check
